+++ components/index.js
@@ -56,6 +56,7 @@ export * from './tbt-form.js';
 export * from './tbt-address.js';
 
 /* Display */
+export * from './tbt-stat.js';
 export * from './tbt-approval-flow.js';
 export * from './tbt-audit-log.js';
 export * from './tbt-timeline.js';